+++ AudioCuesheetEditor/Shared/MainLayout.razor
@@ -90,7 +90,7 @@ along with Foobar.  If not, see
                                             <BarDropdownMenu Background="Background.Dark">
                                                 <BarDropdownItem>
                                                     <Tooltip Placement="TooltipPlacement.Right" Text="@IsCuesheetExportableTooltip">
-                                                        <Button Color="Color.Primary" Clicked="() => ControlModalDialog(modalDownloadCuesheet, true)" Disabled="!IsCuesheetExporable">@_localizer["Download cuesheet"]</Button>
+                                                        <Button Color="Color.Primary" Clicked="() => ControlModalDialog(modalDownloadCuesheet, true)" Disabled="!IsCuesheetExportable">@_localizer["Download cuesheet"]</Button>
                                                     </Tooltip>
                                                 </BarDropdownItem>
                                                 <BarDropdownItem>
@@ -237,7 +237,7 @@ along with Foobar.  If not, see
                 </Validations>
             </ModalBody>
             <ModalFooter>
-                <Button Color="Color.Primary" Clicked="DownloadCuesheetFileClicked" Disabled="!IsCuesheetExporable">@_localizer["Download cuesheet"]</Button>
+                <Button Color="Color.Primary" Clicked="DownloadCuesheetFileClicked" Disabled="!IsCuesheetExportable">@_localizer["Download cuesheet"]</Button>
                 <Button Color="Color.Secondary" Clicked="() => ControlModalDialog(modalDownloadCuesheet, false)">@_localizer["Abort"]</Button>
             </ModalFooter>
         </ModalContent>
@@ -262,7 +262,7 @@ along with Foobar.  If not, see
         get
         {
             String? text = null;
-            if (IsCuesheetExporable == false)
+            if (IsCuesheetExportable == false)
             {
                 text = _localizer["Please check processinghints for errors, otherwise the file is not exportable"];
             }
@@ -270,14 +270,18 @@ along with Foobar.  If not, see
         }
     }
 
-    Boolean IsCuesheetExporable
+    Boolean IsCuesheetExportable
     {
         get
         {
             if (applicationOptions?.Validate(x => x.CuesheetFilename).Status != Model.Entity.ValidationStatus.Error)
             {
-                var generator = new CuesheetfileGenerator(_sessionStateContainer.Cuesheet);
-                return generator.CanWrite;
+                var generator = new ExportfileGenerator(_sessionStateContainer.Cuesheet);
+                if (applicationOptions != null)
+                {
+                    generator.ApplicationOptions = applicationOptions;
+                }
+                return generator.CanWrite(ExportType.Cuesheet);
             }
             else
             {
@@ -452,8 +456,9 @@ along with Foobar.  If not, see
         }
         //Save ApplicationOptions with information from modal popup!
         await _localStorageOptionsProvider.SaveOptions(applicationOptions);
-        var generator = new CuesheetfileGenerator(_sessionStateContainer.Cuesheet);
-        foreach (var result in generator.GenerateCuesheetFiles(applicationOptions.CuesheetFilename))
+        var generator = new ExportfileGenerator(_sessionStateContainer.Cuesheet);
+        generator.ApplicationOptions = applicationOptions;
+        foreach (var result in generator.GenerateExportfiles(ExportType.Cuesheet))
         {
             await _blazorDownloadFileService.DownloadFile(result.Filename, result.Content, "text/plain");
         }
